+++ src/check_datapackage/internals.py
@@ -4,48 +4,6 @@ from typing import Any, Callable, Iterable, TypeVar
 
 from jsonpath import JSONPathMatch, finditer
 
-from check_datapackage.constants import (
-    NAME_PATTERN,
-    PACKAGE_RECOMMENDED_FIELDS,
-    SEMVER_PATTERN,
-)
-
-
-def _add_package_recommendations(schema: dict[str, Any]) -> dict[str, Any]:
-    """Add recommendations from the Data Package standard to the schema.
-
-    Modifies the schema in place.
-
-    Args:
-        schema: The full Data Package schema.
-
-    Returns:
-        The updated Data Package schema.
-    """
-    schema["required"].extend(PACKAGE_RECOMMENDED_FIELDS.keys())
-    schema["properties"]["name"]["pattern"] = NAME_PATTERN
-    schema["properties"]["version"]["pattern"] = SEMVER_PATTERN
-    schema["properties"]["contributors"]["items"]["required"] = ["title"]
-    schema["properties"]["sources"]["items"]["required"] = ["title"]
-    return schema
-
-
-def _add_resource_recommendations(schema: dict[str, Any]) -> dict[str, Any]:
-    """Add recommendations from the Data Resource standard to the schema.
-
-    Modifies the schema in place.
-
-    Args:
-        schema: The full Data Package schema.
-
-    Returns:
-        The updated Data Package schema.
-    """
-    schema["properties"]["resources"]["items"]["properties"]["name"]["pattern"] = (
-        NAME_PATTERN
-    )
-    return schema
-
 
 @dataclass
 class DescriptorField: