+++ tests/test_exclude.py
@@ -5,14 +5,14 @@ from pytest import mark
 from check_datapackage.check import check
 from check_datapackage.config import Config
 from check_datapackage.examples import (
-    example_package_descriptor,
-    example_resource_descriptor,
+    example_package_properties,
+    example_resource_properties,
 )
 from check_datapackage.exclude import Exclude
 
 
 def test_exclude_none_type():
-    descriptor: dict[str, Any] = {
+    properties: dict[str, Any] = {
         "name": "a name",
         "resources": [{"name": "a name", "path": "data.csv"}],
         "contributors": [{"path": "/a/bad/path"}],
@@ -20,18 +20,18 @@ def test_exclude_none_type():
 
     exclude = [Exclude()]
     config = Config(exclude=exclude)
-    issues = check(descriptor, config=config)
+    issues = check(properties, config=config)
 
     assert len(issues) == 1
 
 
 def test_exclude_required_type():
     """Exclude type with the required value."""
-    descriptor = {"name": "a name with spaces"}
+    properties = {"name": "a name with spaces"}
 
     exclude = [Exclude(type="required")]
     config = Config(exclude=exclude)
-    issues = check(descriptor, config=config)
+    issues = check(properties, config=config)
 
     assert len(issues) == 0
 
@@ -39,11 +39,11 @@ def test_exclude_required_type():
 def test_exclude_format_type():
     """Exclude type with the format value."""
     # created must match a date format
-    descriptor = {"created": "20240614"}
+    properties = {"created": "20240614"}
 
     exclude = [Exclude(type="format")]
     config = Config(exclude=exclude)
-    issues = check(descriptor, config=config)
+    issues = check(properties, config=config)
 
     # One issue: missing resources
     assert len(issues) == 1
@@ -51,7 +51,7 @@ def test_exclude_format_type():
 
 def test_exclude_pattern_type():
     """Exclude types with the pattern value."""
-    descriptor: dict[str, Any] = {
+    properties: dict[str, Any] = {
         "name": "a name",
         "resources": [{"name": "a name", "path": "data.csv"}],
         "contributors": [{"path": "/a/bad/path"}],
@@ -59,14 +59,14 @@ def test_exclude_pattern_type():
 
     exclude = [Exclude(type="pattern")]
     config = Config(exclude=exclude)
-    issues = check(descriptor, config=config)
+    issues = check(properties, config=config)
 
     assert len(issues) == 0
 
 
 def test_exclude_multiple_types():
     """Exclude by many types."""
-    descriptor: dict[str, Any] = {
+    properties: dict[str, Any] = {
         "name": "a name",
         "created": "20240614",
         "contributors": [{"path": "/a/bad/path"}],
@@ -78,7 +78,7 @@ def test_exclude_multiple_types():
         Exclude(type="format"),
     ]
     config = Config(exclude=exclude)
-    issues = check(descriptor, config=config)
+    issues = check(properties, config=config)
 
     assert len(issues) == 0
 
@@ -105,60 +105,60 @@ def test_exclude_multiple_types():
     ],
 )
 def test_exclude_jsonpath(jsonpath: str, num_issues: int) -> None:
-    descriptor = example_package_descriptor()
+    properties = example_package_properties()
     # Total 3 issues
-    descriptor["created"] = "20240614"
+    properties["created"] = "20240614"
     # Two issues for resources: type and pattern
-    descriptor["resources"][0]["path"] = "/a/bad/path"
-    descriptor.update({"contributors": [{"path": "/a/bad/path"}]})
+    properties["resources"][0]["path"] = "/a/bad/path"
+    properties.update({"contributors": [{"path": "/a/bad/path"}]})
 
     exclude = [Exclude(jsonpath=jsonpath)]
     config = Config(exclude=exclude)
-    issues = check(descriptor, config=config)
+    issues = check(properties, config=config)
 
     assert len(issues) == num_issues
 
 
 def test_exclude_jsonpath_multiple():
-    descriptor = example_package_descriptor()
-    descriptor["created"] = "20240614"
-    descriptor.update({"contributors": [{"path": "/a/bad/path"}]})
+    properties = example_package_properties()
+    properties["created"] = "20240614"
+    properties.update({"contributors": [{"path": "/a/bad/path"}]})
 
     exclude = [
         Exclude(jsonpath="$.contributors[0].path"),
         Exclude(jsonpath="$.created"),
     ]
     config = Config(exclude=exclude)
-    issues = check(descriptor, config=config)
+    issues = check(properties, config=config)
 
     assert len(issues) == 0
 
 
 def test_exclude_jsonpath_and_type():
-    descriptor = example_package_descriptor()
-    descriptor["contributors"] = [{"path": "/a/bad/path"}, {"path": "/a/bad/path"}]
+    properties = example_package_properties()
+    properties["contributors"] = [{"path": "/a/bad/path"}, {"path": "/a/bad/path"}]
 
     exclude = [
         Exclude(jsonpath="$.contributors[0].path", type="pattern"),
     ]
     config = Config(exclude=exclude)
-    issues = check(descriptor, config=config)
+    issues = check(properties, config=config)
 
     assert len(issues) == 1
 
 
 def test_exclude_jsonpath_and_type_non_overlapping():
-    descriptor = example_package_descriptor()
+    properties = example_package_properties()
     # There should be two issues
-    descriptor["created"] = "20240614"
-    descriptor.update({"contributors": [{"path": "/a/bad/path"}]})
+    properties["created"] = "20240614"
+    properties.update({"contributors": [{"path": "/a/bad/path"}]})
 
     exclude = [
         Exclude(jsonpath="$.contributors[0].path"),
         Exclude(type="pattern"),
     ]
     config = Config(exclude=exclude)
-    issues = check(descriptor, config=config)
+    issues = check(properties, config=config)
 
     # For the created field
     assert len(issues) == 1
@@ -185,8 +185,8 @@ def test_exclude_jsonpath_resources():
     ["$.resources[*].name", "$.resources[1].*"],
 )
 def test_exclude_required_at_jsonpath_array(jsonpath):
-    descriptor = example_package_descriptor()
-    descriptor["resources"].append(example_resource_descriptor())
+    descriptor = example_package_properties()
+    descriptor["resources"].append(example_resource_properties())
     del descriptor["resources"][1]["name"]
 
     exclude = [
@@ -202,7 +202,7 @@ def test_exclude_required_at_jsonpath_array(jsonpath):
     "jsonpath", ["$.*", "$..resources", "..resources", "resources", "..*"]
 )
 def test_exclude_required_at_jsonpath_dict_field(jsonpath):
-    descriptor = example_package_descriptor()
+    descriptor = example_package_properties()
     del descriptor["resources"]
 
     exclude = [